chore: slightly simplified helper function setup - both to ease testing & not purely test dependency code
Changelog: None
